rename test class
